KSENROLL-14960 remove impex artifacts from kuali-student repository
They are being added into the kuali-impex repository.

Remove impex creation logic from the ks-impex/pom.xml

Add in sections to allow the generation of the bundled, app and rice databases
directly from the top level using the ks-impex dependencies.

Needed to change how the artifact version is resolved in
InitAppDbPropertySourceConfig to consider the maven supplied properties.

It used to only look in the environment but with this change will resolve
from the profile properties.