KSENROLL-14743 Reconfigure fusion plugin so that it runs
I had put the fusion plugin configuration into the <pluginManagement> section
of the pom.

However the executions didn't work.  I needed to also configure the plugin
in the <plugins> pom section.

The plugins section version held the executions that controlled the tagging
and the pluginManagement section holds the mappings definition.